requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
Faker==20.1.0
flask-talisman==1.1.0
PyJWT==2.8.0
cryptography==42.0.2
pyOpenSSL==24.0.0
elasticsearch==8.12.0
pyshark==0.6.0
yara-python==4.3.1
psutil==5.9.8
python-evtx==0.7.4
python-registry==1.3.1 